app_celery/requirements.txt
# -*- coding: utf-8 -*-
# Python>=3.11
toollib==1.7.8
python-dotenv==1.1.1
PyYAML==6.0.2
pydantic==2.11.9
celery==5.5.3
redis==6.4.0
gevent==25.9.1